backend/pyproject.toml
[tool.poetry]
name = "promptcraft-backend"
version = "0.1.0"
description = "PromptCraft FastAPI backend"
authors = ["PromptCraft"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.129.0"
uvicorn = "0.40.0"
python-dotenv = "1.2.1"
sqlalchemy = "^2.0.0"
psycopg = "^3.1.0"
transformers = "^4.47.0"
Pillow = "^11.0.0"
psycopg2-binary = "^2.9"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.dependencies.torch]
version = "^2.5.0"
source = "pytorch-cpu"

[tool.poetry.group.dev.dependencies]
pytest = "8.4.2"
httpx = "0.28.1"

[build-system]
requires = ["poetry-core>=1.9.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"
